+++ v2/package.json
@@ -7,6 +7,7 @@
     "@material-ui/icons": "^4.2.1",
     "ace-builds": "^1.3.3",
     "asciidoctor": "^2.0.3",
+    "debug": "^4.1.1",
     "dropbox": "^4.0.28",
     "highlightjs": "^9.12.0",
     "jquery": "^3.3.1",
@@ -35,6 +36,7 @@
   },
   "devDependencies": {
     "@types/ace": "^0.0.42",
+    "@types/debug": "^4.1.4",
     "@types/jest": "^24.0.15",
     "@types/jquery": "^3.3.22",
     "@types/lodash": "^4.14.118",
