+++ v2/public/assets/debug/debug.js
@@ -1,0 +1,1 @@
+../../../node_modules/debug/dist/debug.js
