+++ v2/public/assets/worker/asciidoc-compiler-worker.js
@@ -61,9 +61,9 @@ System.register("compiler", [], function (exports_1, context_1) {
         }
     };
 });
-System.register("asciidoc-compiler", ["compiler"], function (exports_2, context_2) {
+System.register("asciidoc-compile", ["compiler"], function (exports_2, context_2) {
     "use strict";
-    var compiler_1, asciidoctor, BlockingAsciidocCompiler, AsyncAsciidocCompiler;
+    var compiler_1, asciidoctor;
     var __moduleName = context_2 && context_2.id;
     function asciidocCompile(request) {
         var isInline = request.outputType == compiler_1.OutputType.PREVIEW;
@@ -103,7 +103,7 @@ System.register("asciidoc-compiler", ["compiler"], function (exports_2, context_
             elapsedTimeMs: new Date().getTime() - startTs.getTime()
         };
     }
-    exports_2("asciidocCompile", asciidocCompile);
+    exports_2("default", asciidocCompile);
     return {
         setters: [
             function (compiler_1_1) {
@@ -111,78 +111,32 @@ System.register("asciidoc-compiler", ["compiler"], function (exports_2, context_
             }
         ],
         execute: function () {
-            BlockingAsciidocCompiler = /** @class */ (function (_super) {
-                __extends(BlockingAsciidocCompiler, _super);
-                function BlockingAsciidocCompiler() {
-                    return _super !== null && _super.apply(this, arguments) || this;
-                }
-                BlockingAsciidocCompiler.prototype.compile = function (request) {
-                    return Promise.resolve(asciidocCompile(request));
-                };
-                return BlockingAsciidocCompiler;
-            }(compiler_1.Compiler));
-            exports_2("BlockingAsciidocCompiler", BlockingAsciidocCompiler);
-            AsyncAsciidocCompiler = /** @class */ (function (_super) {
-                __extends(AsyncAsciidocCompiler, _super);
-                function AsyncAsciidocCompiler() {
-                    var _this = _super.call(this) || this;
-                    _this.worker = new Worker('/assets/worker/asciidoc-compiler-worker-loader.js');
-                    _this.pendingRequests = new Map();
-                    _this.worker.addEventListener('message', _this.onMessage.bind(_this));
-                    return _this;
-                }
-                AsyncAsciidocCompiler.prototype.compile = function (request) {
-                    var _this = this;
-                    return new Promise(function (resolve) {
-                        _this.pendingRequests.set(request.requestId, resolve);
-                        _this.log("Issuing request with requestId " + request.requestId);
-                        _this.worker.postMessage(request);
-                    });
-                };
-                AsyncAsciidocCompiler.prototype.onMessage = function (ev) {
-                    var result = ev.data;
-                    if (!this.pendingRequests.has(result.requestId)) {
-                        this.log("Discarding result for unexpected requestId " + result.requestId);
-                        return;
-                    }
-                    this.log("Received result for requestId " + result.requestId);
-                    this.pendingRequests.get(result.requestId)(result);
-                    this.pendingRequests["delete"](result.requestId);
-                };
-                AsyncAsciidocCompiler.prototype.log = function (message) {
-                    console.log("[AsyncAsciidocCompiler] " + message);
-                };
-                return AsyncAsciidocCompiler;
-            }(compiler_1.Compiler));
-            exports_2("AsyncAsciidocCompiler", AsyncAsciidocCompiler);
         }
     };
 });
-System.register("asciidoc-compiler-worker", ["asciidoc-compiler"], function (exports_3, context_3) {
+System.register("asciidoc-compiler-worker", ["asciidoc-compile"], function (exports_3, context_3) {
     "use strict";
-    var asciidoc_compiler_1, AsciidocCompilerWorker;
+    var asciidoc_compile_1, AsciidocCompilerWorker;
     var __moduleName = context_3 && context_3.id;
     return {
         setters: [
-            function (asciidoc_compiler_1_1) {
-                asciidoc_compiler_1 = asciidoc_compiler_1_1;
+            function (asciidoc_compile_1_1) {
+                asciidoc_compile_1 = asciidoc_compile_1_1;
             }
         ],
         execute: function () {
             AsciidocCompilerWorker = /** @class */ (function () {
                 function AsciidocCompilerWorker() {
+                    this.log = debug('AsciidocCompilerWorker');
                     this.log('Starting Asciidoc compile worker');
                     self.addEventListener('message', this.onMessage.bind(this));
                 }
                 AsciidocCompilerWorker.prototype.onMessage = function (ev) {
                     this.log('Received compile request');
                     var request = ev.data;
-                    var result = asciidoc_compiler_1.asciidocCompile(request);
+                    var result = asciidoc_compile_1["default"](request);
                     this.log('Completed compile request');
                     self.postMessage(result);
-                };
-                AsciidocCompilerWorker.prototype.log = function (message) {
-                    self.console && self.console.info("[AsciidocCompilerWorker] " + message);
                 };
                 return AsciidocCompilerWorker;
             }());
