+++ v2/src/document/asciidoc-compiler.ts
@@ -1,64 +1,6 @@
-import {Compiler, CompileRequest, CompileResult, OutputType} from './compiler';
-
-declare var Opal: any;
-declare var Asciidoctor: any;
-declare var html_beautify: any;
-declare var hljs: any;
-
-let asciidoctor: {
-  convert: (body: string, options?: any) => string;
-};
-
-export function asciidocCompile(request: CompileRequest): CompileResult {
-  const isInline = request.outputType == OutputType.PREVIEW;
-  const shouldBeautify =
-    request.outputType == OutputType.DISPLAY_HTML ||
-    request.outputType == OutputType.EXPORT_HTML;
-  const beautifyOptions = {
-    indent_size: 2,
-    wrap_line_length: 80,
-  };
-  const shouldHighlight = request.outputType == OutputType.DISPLAY_HTML;
-
-  if (asciidoctor === undefined) {
-    asciidoctor = Asciidoctor();
-  }
-
-  const startTs = new Date();
-  let compiledBody;
-  if (isInline) {
-    compiledBody = asciidoctor.convert(
-      request.body,
-      Opal.hash2(['attributes'], {
-        attributes: ['showtitle'],
-      })
-    );
-  } else {
-    const cssPath = '/assets/asciidoctor/css/asciidoctor.css';
-    compiledBody = asciidoctor.convert(
-      request.body,
-      Opal.hash2(['header_footer', 'attributes'], {
-        header_footer: true,
-        attributes: ['nofooter', 'copycss', 'stylesheet=' + cssPath],
-      })
-    );
-  }
-  if (shouldBeautify) {
-    compiledBody = html_beautify(compiledBody, beautifyOptions);
-  }
-  if (shouldHighlight) {
-    compiledBody = hljs.highlight(
-      'html',
-      compiledBody,
-      true /* allow malformed input */
-    ).value;
-  }
-  return {
-    requestId: request.requestId,
-    compiledBody,
-    elapsedTimeMs: new Date().getTime() - startTs.getTime(),
-  };
-}
+import debug from 'debug';
+import asciidocCompile from './asciidoc-compile';
+import {Compiler, CompileRequest, CompileResult} from './compiler';
 
 export class BlockingAsciidocCompiler extends Compiler {
   compile(request: CompileRequest) {
@@ -93,10 +35,7 @@ export class AsyncAsciidocCompiler extends Compiler {
     this.pendingRequests.delete(result.requestId);
   }
 
-  private log(message: string) {
-    console.log(`[AsyncAsciidocCompiler] ${message}`);
-  }
-
+  private readonly log = debug('AsyncAsciidocCompiler');
   private readonly worker = new Worker(
     '/assets/worker/asciidoc-compiler-worker-loader.js'
   );
